refactor(monorepo): move `getBundleBanner` to `utils/` directory
