+++ scripts/utils/getBundleBanner.mjs
